Code refactor.
