+++ packages/Webkul/Shop/src/Http/Controllers/API/OnepageController.php
@@ -1,0 +1,250 @@
+<?php
+
+namespace Webkul\Shop\Http\Controllers\API;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+use Webkul\Shop\Http\Resources\CartResource;
+use Webkul\Shop\Http\Controllers\API\APIController;
+use Webkul\Shipping\Facades\Shipping;
+use Webkul\Sales\Repositories\OrderRepository;
+use Webkul\Payment\Facades\Payment;
+use Webkul\Customer\Repositories\CustomerRepository;
+use Webkul\Checkout\Http\Requests\CustomerAddressForm;
+use Webkul\Checkout\Facades\Cart;
+
+class OnepageController extends APIController
+{
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct(
+        protected OrderRepository $orderRepository,
+        protected CustomerRepository $customerRepository
+    )
+    {
+        parent::__construct();
+    }
+
+    /**
+     * Return order short summary.
+     */
+    public function summary(): JsonResource
+    {
+        $cart = Cart::getCart();
+
+        return new CartResource($cart);
+    }
+
+    /**
+     * Store customer address.
+     *
+     * @param  \Webkul\Checkout\Http\Requests\CustomerAddressForm  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function storeAddress(CustomerAddressForm $request): JsonResource
+    {
+        $data = $request->all();
+
+        if (
+            ! auth()->guard('customer')->check()
+            && ! Cart::getCart()->hasGuestCheckoutItems()
+        ) {
+            return new JsonResource([
+                'redirect' => true,
+                'data'     => route('shop.customer.session.index')
+            ]);
+        }
+
+        $data['billing']['address1'] = implode(PHP_EOL, array_filter($data['billing']['address1']));
+        
+        $data['shipping']['address1'] = implode(PHP_EOL, array_filter($data['shipping']['address1'])); 
+
+        if (
+            Cart::hasError()
+            || ! Cart::saveCustomerAddress($data)
+        ) {
+            return new JsonResource([
+                'redirect' => true,
+                'data'     => route('shop.checkout.cart.index')
+            ]);
+        }
+
+        $cart = Cart::getCart();
+
+        Cart::collectTotals();
+
+        if ($cart->haveStockableItems()) {
+            if (! $rates = Shipping::collectRates()) {
+                return new JsonResource([
+                    'redirect' => true,
+                    'data'     => route('shop.checkout.cart.index')
+                ]);
+            }
+
+            return new JsonResource([
+                'redirect' => false,
+                'data'     => $rates
+            ]);
+        }
+
+        return new JsonResource([
+            'redirect' => false,
+            'data'     => Payment::getSupportedPaymentMethods()
+        ]);
+    }
+
+    /**
+     * Validate order before creation.
+     *
+     * @return void|\Exception
+     */
+    public function validateOrder()
+    {
+        $cart = Cart::getCart();
+
+        $minimumOrderAmount = core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?: 0;
+
+        if (
+            auth()->guard('customer')->check()
+            && auth()->guard('customer')->user()->is_suspended
+        ) {
+            throw new \Exception(trans('shop::app.checkout.cart.suspended-account-message'));
+        }
+
+        if (
+            auth()->guard('customer')->user()
+            && ! auth()->guard('customer')->user()->status
+        ) {
+            throw new \Exception(trans('shop::app.checkout.cart.inactive-account-message'));
+        }
+
+        if (! $cart->checkMinimumOrder()) {
+            throw new \Exception(trans('shop::app.checkout.cart.minimum-order-message', ['amount' => core()->currency($minimumOrderAmount)]));
+        }
+
+        if ($cart->haveStockableItems() && ! $cart->shipping_address) {
+            throw new \Exception(trans('shop::app.checkout.cart.check-shipping-address'));
+        }
+
+        if (! $cart->billing_address) {
+            throw new \Exception(trans('shop::app.checkout.cart.check-billing-address'));
+        }
+
+        if (
+            $cart->haveStockableItems()
+            && ! $cart->selected_shipping_rate
+        ) {
+            throw new \Exception(trans('shop::app.checkout.cart.specify-shipping-method'));
+        }
+
+        if (! $cart->payment) {
+            throw new \Exception(trans('shop::app.checkout.cart.specify-payment-method'));
+        }
+    }
+
+    /**
+     * Store order
+     */
+    public function storeOrder(): JsonResource
+    {
+        if (Cart::hasError()) {
+            return new JsonResource([
+                'redirect'     => true,
+                'redirect_url' => route('shop.checkout.cart.index')
+            ]);
+        }
+
+        Cart::collectTotals();
+
+        $this->validateOrder();
+
+        $cart = Cart::getCart();
+
+        if ($redirectUrl = Payment::getRedirectUrl($cart)) {
+            return new JsonResource([
+                'redirect'     => true,
+                'redirect_url' => $redirectUrl
+            ]);
+        }
+
+        $order = $this->orderRepository->create(Cart::prepareDataForOrder());
+
+        Cart::deActivateCart();
+
+        Cart::activateCartIfSessionHasDeactivatedCartId();
+
+        session()->flash('order', $order);
+
+        return new JsonResource([
+            'redirect'     => true,
+            'redirect_url' => route('shop.checkout.onepage.success')
+        ]);
+    }
+
+    /**
+     * Store shipping method.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function storeShippingMethod()
+    {
+        $shippingMethod = request()->get('shipping_method');
+
+        if (
+            Cart::hasError()
+            || ! $shippingMethod
+            || ! Cart::saveShippingMethod($shippingMethod)
+        ) {
+            return response()->json(['redirect_url' => route('shop.checkout.cart.index')], 403);
+        }
+
+        Cart::collectTotals();
+
+        return response()->json(Payment::getSupportedPaymentMethods());
+    }
+
+    /**
+     * Store payment method.
+     *
+     * @return array
+     */
+    public function storePaymentMethod()
+    {
+        $payment = request()->get('payment');
+
+        if (
+            Cart::hasError()
+            || ! $payment
+            || ! Cart::savePaymentMethod($payment)
+        ) {
+            return response()->json(['redirect_url' => route('shop.checkout.cart.index')], 403);
+        }
+
+        Cart::collectTotals();
+
+        $cart = Cart::getCart();
+
+        return [
+            'cart' => new CartResource($cart),
+        ];
+    }
+
+    /**
+     * Check for minimum order.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function checkMinimumOrder()
+    {
+        $minimumOrderAmount = (float) core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?: 0;
+
+        $status = Cart::checkMinimumOrder();
+
+        return response()->json([
+            'status'  => ! $status ? false : true,
+            'message' => ! $status ? trans('shop::app.checkout.cart.minimum-order-message', ['amount' => core()->currency($minimumOrderAmount)]) : 'Success',
+        ]);
+    }
+}
