+++ packages/Webkul/Shop/src/Http/Resources/CartResource.php
@@ -15,6 +15,10 @@ class CartResource extends JsonResource
      */
     public function toArray($request)
     {
+        $taxes = collect(Tax::getTaxRatesWithAmount($this, true))->map(function ($rate) {
+            return core()->currency($rate ?? 0);
+        });
+
         return [
             'id'                             => $this->id,
             'items_count'                    => $this->items_count,
@@ -22,7 +26,7 @@ class CartResource extends JsonResource
             'grand_total'                    => $this->grand_total,
             'base_sub_total'                 => core()->currency($this->base_sub_total),
             'base_tax_total'                 => $this->base_tax_total,
-            'base_tax_amounts'               => collect(Tax::getTaxRatesWithAmount($this, true)),
+            'base_tax_amounts'               => $taxes,
             'formatted_base_discount_amount' => core()->currency($this->base_discount_amount),
             'base_discount_amount'           => $this->base_discount_amount,
             'base_grand_total'               => core()->currency($this->base_grand_total),
