+++ packages/Webkul/Shop/src/Resources/views/checkout/onepage/index.blade.php
@@ -1,4 +1,5 @@
 <x-shop::layouts
+    :has-header="false"
     :has-feature="false"
     :has-footer="false"
 >
