+++ packages/Webkul/Shop/src/Resources/views/checkout/onepage/summary.blade.php
@@ -1,10 +1,12 @@
 {!! view_render_event('bagisto.shop.checkout.cart.summary.before') !!}
+
 <v-cart-summary
     ref="vCartSummary"
     :cart="cart"
     :is-cart-loading="isCartLoading"
 >
 </v-cart-summary>
+
 {!! view_render_event('bagisto.shop.checkout.cart.summary.after') !!}
 
 @pushOnce('scripts')
@@ -16,7 +18,9 @@
         <template v-else>
             <div class="max-w-full w-[442px] pl-[30px] h-max sticky top-[30px] max-lg:w-auto max-lg:max-w-[442px] max-lg:pl-0">
 
-                <h2 class="text-[26px] font-medium max-sm:text-[20px]">@lang('Cart Summary')</h2>
+                <h2 class="text-[26px] font-medium max-sm:text-[20px]">
+                    @lang('Cart Summary')
+                </h2>
                 
                 <div class="grid border-b-[1px] border-[#E9E9E9] mt-[40px] max-sm:mt-[20px]">
                     <div 
@@ -29,17 +33,32 @@
                             :title="item.name"
                             :alt="item.name"
                         />
-                        <div class="">
-                            <p class="text-[16px] max-sm:text-[14px] max-sm:font-medium text-navyBlue">@{{ item.name }}</p>
-                            <p class="text-[18px] font-medium mt-[10px] max-sm:text-[14px] max-sm:font-normal">@{{ item.formatted_total }} X @{{ item.quantity }}</p>
+
+                        <div>
+                            <p 
+                                class="text-[16px] max-sm:text-[14px] max-sm:font-medium text-navyBlue" 
+                                v-text="item.name"
+                            >
+                            </p>
+
+                            <p class="text-[18px] font-medium mt-[10px] max-sm:text-[14px] max-sm:font-normal">
+                                @{{ item.formatted_total }} X @{{ item.quantity }}
+                            </p>
                         </div>
                     </div>
                 </div>
 
                 <div class="grid gap-[15px] mt-[25px] mb-[30px]">
                     <div class="flex text-right justify-between">
-                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-normal">@lang('Subtotal')</p>
-                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-medium font-medium">@{{ cart.base_sub_total  }}</p>
+                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-normal">
+                            @lang('Subtotal')
+                        </p>
+
+                        <p 
+                            class="text-[16px] max-sm:text-[14px] max-sm:font-medium font-medium"
+                            v-text="cart.base_sub_total"
+                        >
+                        </p>
                     </div>
 
                     <div 
@@ -47,41 +66,71 @@
                         v-for="(amount, index) in cart.base_tax_amounts"
                         v-if="parseFloat(cart.base_tax_total)"
                     >
-                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-normal">Tax (@{{ index }})%</p>
-                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-medium font-medium">@{{ amount }}</p>
+                        <p class="text-[16px] max-sm:text-[14px] max-sm:font-normal">
+                            @lang('Tax') (@{{ index }})%
+                        </p>
+
+                        <p 
+                            class="text-[16px] max-sm:text-[14px] max-sm:font-medium font-medium"
+                            v-text="amount"
+                        >
+                        </p>
                     </div>
 
                     <div 
                         class="flex text-right justify-between"
                         v-if="cart.selected_shipping_rate"
                     >
-                        <p class="text-[16px]">@lang('Delivery Charges')</p>
-                        <p class="text-[16px] font-medium">@{{ cart.selected_shipping_rate }}</p>
+                        <p class="text-[16px]">
+                            @lang('Delivery Charges')
+                        </p>
+
+                        <p 
+                            class="text-[16px] font-medium"
+                            v-text="cart.selected_shipping_rate"
+                        >
+                        </p>
                     </div>
 
                     <div 
                         class="flex text-right justify-between"
                         v-if="cart.base_discount_amount && parseFloat(cart.base_discount_amount) > 0"
                     >
-                        <p class="text-[16px]">@lang('Discount amount')</p>
-                        <p class="text-[16px] font-medium">@{{ cart.formatted_base_discount_amount }}</p>
+                        <p class="text-[16px]">
+                            @lang('Discount amount')
+                        </p>
+
+                        <p 
+                            class="text-[16px] font-medium"
+                            v-text="cart.formatted_base_discount_amount"
+                        >
+                        </p>
                     </div>
 
                     @include('shop::checkout.onepage.coupon')
 
                     <div class="flex text-right justify-between">
-                        <p class="text-[18px] font-semibold">@lang('Grand total')</p>
-                        <p class="text-[18px] font-semibold"> @{{ cart.base_grand_total }}</p>
+                        <p class="text-[18px] font-semibold">
+                            @lang('Grand total')
+                        </p>
+
+                        <p 
+                            class="text-[18px] font-semibold"
+                            v-text="cart.base_grand_total"
+                        >
+                        </p>
                     </div>
                 </div>
 
-                <button
-                    class="block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer max-sm:text-[14px] max-sm:px-[25px] max-sm:mb-[40px]"
-                    v-if="isPlaceOrderVisible"
-                    @click="placeOrder"
-                >
-                    @lang('Place order')    
-                </button>
+                <div class="flex justify-end">
+                    <button
+                        class="block bg-navyBlue text-white text-base w-max font-medium py-[11px] px-[43px] rounded-[18px] text-center cursor-pointer max-sm:text-[14px] max-sm:px-[25px] max-sm:mb-[40px]"
+                        v-if="isPlaceOrderVisible"
+                        @click="placeOrder"
+                    >
+                        @lang('Place order')    
+                    </button>
+                </div>
             </div>
         </template>
     </script>
@@ -104,12 +153,10 @@
                             '_token': "{{ csrf_token() }}"
                         })
                         .then(response => {
-                            if (response.data.success) {
-                                if (response.data.redirect_url) {
-                                    window.location.href = response.data.redirect_url;
-                                } else {
-                                    window.location.href = "{{ route('shop.checkout.onepage.success') }}";
-                                }
+                            if (response.data.data.redirect) {
+                                window.location.href = response.data.data.redirect_url;
+                            } else {
+                                window.location.href = "{{ route('shop.checkout.onepage.success') }}";
                             }
                         })
                         .catch(error => console.log(error))
